Restructure dependencies
